requirements.txt
requests
feedparser
instagrapi
SpeechRecognition
huggingface_hub
fastapi
uvicorn
piper-tts
python-multipart
python-dotenv
pydub
keyring
pathvalidate